feat(linter): generalize no_elvis_operator into no_shorthand_ternary
