+++ crates/linter/src/plugin/strictness/rules/mod.rs
@@ -1,7 +1,7 @@
 pub mod missing_assert_description;
 pub mod no_assignment_in_condition;
-pub mod no_elvis_operator;
 pub mod no_empty_construct;
+pub mod no_shorthand_ternary;
 pub mod require_constant_type;
 pub mod require_identity_comparison;
 pub mod require_parameter_type;
